pyproject.toml
[project]
name = "telegram-suggestion-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot",
    "python-dotenv"
]
